backend/requirements.txt
fastapi
uvicorn
pydantic
requests
beautifulsoup4
openai
python-dotenv
chromadb
tiktoken
duckduckgo-search
